Update bundle_pl.properties
I filled in the blanks by the previous translator. They left some descriptions and item names unchanged or with a placeholder. I decided to rename Spectre from Huragan (hurricane) to Kostucha (slavic death reaper character in folklore) to keep up with the paranormal theme. And renamed a few unit names to fit the theme too